+++ .gitignore
@@ -49,7 +49,8 @@ coverage.xml
 # Sphinx documentation
 docs/_build/
 
-# JetBrains and Notebooks
+# JetBrains and Notebooks and more
 .idea/
 .ipynb_checkpoints/
+.UI
 Scratch.ipynb
