+++ setup.py
@@ -1,0 +1,12 @@
+from distutils.core import setup
+
+setup(
+    name='PyCCD-Plotting-Tool',
+    version='1.0.0',
+    packages=['Controls', 'PlotFrame', 'Plotting', 'RetrieveData'],
+    url='https://github.com/danzelenak-usgs/PyCCD-Plotting-Tool',
+    license='',
+    author='Daniel Zelenak',
+    author_email='daniel.zelenak.ctr@usgs.gov',
+    description='Plotting tool for PyCCD results and ARD observations'
+)
